refactor: separate out tutorial view
